requirements.txt
asgiref==3.7.2
Django==3.2.23
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
greenlet==3.0.3
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
Mako==1.2.4
MarkupSafe==2.1.4
mysqlclient==2.1.1
pytz==2023.3.post1
SQLAlchemy==2.0.25
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.15.0
